layout/style/test/test_extra_inherit_initial.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=940229
-->
<head>
  <title>Test handling extra inherit/initial/unset in CSS declarations (Bug 940229)</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="property_database.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=940229">Mozilla Bug 940229</a>
<p id="display"></p>
<div id="content" style="display: none">

<div id="testnode"></div>
  
</div>
<pre id="test">
<script class="testbody" type="text/javascript">

/*
 * Inspired by mistake in quotes noticed while reviewing bug 189519.
 */

var gPropsNeedComma = {
  "font": true,
  "font-family": true,
  "voice-family": true,
};

var gElement = document.getElementById("testnode");
var gDeclaration = gElement.style;

function test_property(property)
{
  var info = gCSSProperties[property];

  var delim = (property in gPropsNeedComma) ? ", " : " ";

  function test_value_pair(relation, val1, val2, extraval) {
    var decl = property + ": " + val1 + delim + val2;
    gElement.setAttribute("style", decl);
    if ("subproperties" in info) {
      for (idx in info.subproperties) {
        var subprop = info.subproperties[idx];
        is(gDeclaration.getPropertyValue(subprop), "",
           ["expected", extraval, "ignored", relation, "value in",
            "'" + decl + "'", "when looking at subproperty",
            "'" + subprop + "'"].join(" "));
      }
    } else {
      is(gDeclaration.getPropertyValue(property), "",
         ["expected", extraval, "ignored", relation, "value in",
          "'" + decl + "'"].join(" "));
    }
  }

  function test_value(value) {
    test_value_pair("after", value, "inherit", "inherit");
    test_value_pair("after", value, "initial", "initial");
    test_value_pair("after", value, "unset", "unset");
    test_value_pair("before", "inherit", value, "inherit");
    test_value_pair("before", "initial", value, "initial");
    test_value_pair("before", "unset", value, "unset");
  }

  var idx;
  for (idx in info.initial_values)
    test_value(info.initial_values[idx]);
  for (idx in info.other_values)
    test_value(info.other_values[idx]);
}

SimpleTest.waitForExplicitFinish();
SimpleTest.requestLongerTimeout(4);

function start_test() {
  for (var prop in gCSSProperties) {
    test_property(prop);
  }
  SimpleTest.finish();
}

// Turn off CSS error reporting for this test, since it's a bit expensive,
// and we're expecting to generate tons and tons of parse errors here.
SpecialPowers.pushPrefEnv({ "set": [["layout.css.report_errors", false]] },
                          start_test);

</script>
</pre>
</body>
</html>
